requirements.txt
google-cloud-vision==3.10.2
python-dotenv==1.0.1
playwright==1.51.0
PyMuPDF==1.25.3
requests==2.32.5
certifi==2025.1.31
pytest==8.3.5
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
pymongo==4.10.1
python-dateutil==2.9.0.post0
rapidfuzz==3.12.2
